Replace the original discovery with the new version and re-enable tests
